+++ packages/amplify-cli-core/src/index.ts
@@ -1,3 +1,4 @@
+import { ViewResourceTableParams } from './cliViewAPI';
 import { ServiceSelection } from './serviceSelection';
 
 export * from './cfnUtilities';
@@ -21,6 +22,7 @@ export * from './utils';
 export * from './banner-message';
 export * from './cliGetCategories';
 export * from './cliRemoveResourcePrompt';
+export * from "./cliViewAPI";
 
 // Temporary types until we can finish full type definition across the whole CLI
 
@@ -30,6 +32,9 @@ export type $TSAny = any;
 // Use it for all CLI Context class references, it enables a quick way to see what we have on the context
 export type $TSContext = {
   amplify: AmplifyToolkit;
+  /**
+   * @deprecated Use printer from package amplify-prompts instead
+   */
   print: IContextPrint;
   migrationInfo: $TSAny;
   projectHasMobileHubResources: boolean;
@@ -49,16 +54,49 @@ export type CategoryName = string;
 export type ResourceName = string;
 
 export type IContextPrint = {
+  /**
+   * @deprecated Use printer.info from amplify-prommpts instead
+   */
   info: (message: string) => void;
+  /**
+   * @deprecated Why are you using this? If you really need it, implement it in amplify-prompts printer.ts
+   */
   fancy: (message?: string) => void;
+  /**
+   * @deprecated Use printer.warn from amplify-prompts instead
+   */
   warning: (message: string) => void;
+  /**
+   * @deprecated Use printer.error from amplify-prompts instead
+   */
   error: (message: string) => void;
+  /**
+   * @deprecated Use printer.success from amplify-prommpts instead
+   */
   success: (message: string) => void;
+  /**
+   * @deprecated The next time we refactor code that uses this function, refactor the table function into formatter.ts from amplify-prompts and use that instead
+   */
   table: (data: string[][], options?: { format?: 'markdown' | 'lean' }) => void;
+  /**
+   * @deprecated Use printer.debug from amplify-prompts instead
+   */
   debug: (message: string) => void;
+  /**
+   * @deprecated Use printer.info from amplify-prompts and specify color
+   */
   green: (message: string) => void;
+  /**
+   * @deprecated Use printer.info from amplify-prompts and specify color
+   */
   yellow: (message: string) => void;
+  /**
+   * @deprecated Use printer.info from amplify-prompts and specify color
+   */
   red: (message: string) => void;
+  /**
+   * @deprecated Use printer.info from amplify-prompts and specify color
+   */
   blue: (message: string) => void;
 };
 
@@ -164,6 +202,7 @@ export interface AmplifyProjectConfig {
 
 export type $TSCopyJob = any;
 
+
 // Temporary interface until Context refactor
 interface AmplifyToolkit {
   confirmPrompt: (prompt: string, defaultValue?: boolean) => Promise<boolean>;
@@ -188,6 +227,9 @@ interface AmplifyToolkit {
   getResourceStatus: (category?: $TSAny, resourceName?: $TSAny, providerName?: $TSAny, filteredResources?: $TSAny) => $TSAny;
   getResourceOutputs: () => $TSAny;
   getWhen: () => $TSAny;
+  /**
+   * @deprecated Use validators from amplify-prompts or add a new validator in that module
+   */
   inputValidation: (input: $TSAny) => (value: $TSAny) => boolean | string;
   listCategories: () => $TSAny;
   makeId: (n?: number) => string;
@@ -214,8 +256,9 @@ interface AmplifyToolkit {
   sharedQuestions: () => $TSAny;
   showAllHelp: () => $TSAny;
   showHelp: (header: string, commands: { name: string; description: string }[]) => $TSAny;
-  showHelpfulProviderLinks: () => $TSAny;
+  showHelpfulProviderLinks: (context : $TSContext) => $TSAny;
   showResourceTable: () => $TSAny;
+  showStatusTable:( resourceTableParams : ViewResourceTableParams )=> $TSAny; //Enhanced Status with CFN-Diff
   serviceSelectionPrompt: (
     context: $TSContext,
     category: string,