+++ packages/amplify-cli/src/extensions/amplify-helpers/confirm-prompt.ts
@@ -1,5 +1,8 @@
 import * as inquirer from 'inquirer';
 
+/**
+ * @deprecated Use confirmContinue from ammplify-prompts instead
+ */
 export async function confirmPrompt(message: string, defaultValue = true) {
   const ans = await inquirer.prompt({
     name: 'yesno',