+++ packages/amplify-graphql-index-transformer/src/index.ts
@@ -1,1 +1,2 @@
+export { IndexTransformer } from './graphql-index-transformer';
 export { PrimaryKeyTransformer } from './graphql-primary-key-transformer';