+++ packages/amplify-graphql-transformer-core/src/index.ts
@@ -15,6 +15,7 @@ export {
   LogConfig,
 } from './transformation';
 export { DeploymentResources } from './transformation/types';
+export { validateModelSchema } from './transformation/validation';
 export {
   ConflictDetectionType,
   ConflictHandlerType,