+++ packages/amplify-headless-interface/src/interface/api/add.ts
@@ -166,7 +166,7 @@ export interface AppSyncCognitoUserPoolsAuthType {
   /**
    * The user pool that will be used to authenticate requests.
    */
-  cognitoUserPoolId: string;
+  cognitoUserPoolId?: string;
 }
 
 /**